requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
Pillow==9.3.0
requests==2.28.1
Shapely==1.8.0
staticmap==0.5.5
urllib3==1.26.13
azure-functions